优化表结构

+++ cloud-base-service/cloud-user/src/main/java/org/cloud/user/param/ServeTableParam.java
@@ -11,27 +11,27 @@ import lombok.Data;
 @Schema(name = "同步表响应对象", description = "同步表响应对象")
 public class ServeTableParam {
 
-	/**
+    /**
      * id
      */
     @Schema(description = "id")
     private Long id;
 
-	/**
+    /**
      * 采集任务Id
      */
     @Schema(description = "采集任务Id")
     private Long serveId;
 
-	/**
+    /**
      * 数据表映射
      */
     @Schema(description = "数据表映射")
     private Long tableMapId;
 
-	/**
+    /**
      * 版本
      */
     @Schema(description = "版本")
-    private Long version;
+    private Integer version;
 }
